# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NCC-ERP-AMS API Documentation",
        "description": "API Documentation for NCC-ERP-AMS System",
        "contact": {
            "name": "NCC Admin",
            "url": "https://ncc.asia",
            "email": "it@ncc.asia"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/webhooks": {
            "get": {
                "tags": [
                    "Webhook"
                ],
                "summary": "Get list of webhooks",
                "description": "Returns list of registered webhooks",
                "operationId": "getWebhooks",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter results by name or URL",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "webhook_id",
                        "in": "query",
                        "description": "Limit results to a specific webhook ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset to start results from",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of results to return",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Column to sort results by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "name",
                                "url",
                                "created_at",
                                "updated_at",
                                "type"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "deleted",
                        "in": "query",
                        "description": "Include deleted webhooks",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WebhooksResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Webhook"
                ],
                "summary": "Create a new webhook",
                "description": "Creates a new webhook and returns the newly created webhook",
                "operationId": "createWebhook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "url",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Asset Created Hook"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://example.com/webhook/asset-created"
                                    },
                                    "type": {
                                        "description": "Array of webhook event types this webhook subscribes to",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "CHECKOUT_ASSET"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Webhook was created successfully"
                                        },
                                        "payload": {
                                            "$ref": "#/components/schemas/Webhook"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/webhooks/{webhook}": {
            "get": {
                "tags": [
                    "Webhook"
                ],
                "summary": "Get a single webhook",
                "description": "Returns a single webhook",
                "operationId": "getWebhookById",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "description": "ID of webhook to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Webhook"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Webhook not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Webhook"
                ],
                "summary": "Update a webhook",
                "description": "Updates an existing webhook",
                "operationId": "updateWebhook",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "description": "ID of webhook to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Asset Updated Hook"
                                    },
                                    "url": {
                                        "type": "string",
                                        "example": "https://example.com/webhook/asset-updated"
                                    },
                                    "type": {
                                        "description": "Array of webhook event types this webhook subscribes to",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "CHECKOUT_ASSET"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Webhook updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Webhook was updated successfully"
                                        },
                                        "payload": {
                                            "$ref": "#/components/schemas/Webhook"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Webhook not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Webhook"
                ],
                "summary": "Delete a webhook",
                "description": "Deletes an existing webhook",
                "operationId": "deleteWebhook",
                "parameters": [
                    {
                        "name": "webhook",
                        "in": "path",
                        "description": "ID of webhook to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "messages": {
                                            "type": "string",
                                            "example": "Webhook was deleted successfully"
                                        },
                                        "payload": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Webhook not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DateObject": {
                "properties": {
                    "date": {
                        "type": "string",
                        "example": "2024-01-15 10:30:00"
                    },
                    "formatted": {
                        "type": "string",
                        "example": "Jan 15, 2024 10:30 AM"
                    }
                },
                "type": "object"
            },
...
            "AvailableWebhookActions": {
                "properties": {
                    "update": {
                        "type": "boolean",
                        "example": true
                    },
                    "restore": {
                        "type": "boolean",
                        "example": false
                    },
                    "delete": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "Webhook": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Asset Checkout Webhook"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://webhook.mezon.ai/webhooks/1957415919202340864/MTc1NTQ5MDYyODk0..."
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/DateObject"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/DateObject"
                    },
                    "type": {
                        "description": "Array of webhook event types this webhook subscribes to",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "CHECKOUT_TOOL"
                        }
                    },
                    "available_actions": {
                        "$ref": "#/components/schemas/AvailableWebhookActions"
                    }
                },
                "type": "object"
            },
            "WebhooksResponse": {
                "description": "Webhooks response format",
                "properties": {
                    "total": {
                        "description": "Total number of webhooks",
                        "type": "integer",
                        "example": 10
                    },
                    "rows": {
                        "description": "Array of webhook items",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Webhook"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter Bearer token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}